test: start transaction again, when container reboot comitted it, move tests to unit
