+++ tests/integration/Core/Framework/DataAbstractionLayer/VersionManagerTest.php
@@ -73,6 +73,7 @@ class VersionManagerTest extends TestCase
             ALTER TABLE `product`
             DROP FOREIGN KEY `fk.product.many_to_one_id`;
         ');
+        $this->connection->executeStatement('DROP TABLE IF EXISTS `many_to_one_product`');
         $this->connection->executeStatement('DROP TABLE `many_to_one_product`');
         $this->connection->executeStatement('
             ALTER TABLE `product`
