+++ tests/unit/Core/Framework/App/Event/AppActivatedEventTest.php
@@ -1,31 +1,33 @@
 <?php declare(strict_types=1);
 
-namespace Cicada\Tests\Integration\Core\Framework\App\Event;
+namespace Cicada\Tests\Unit\Core\Framework\App\Event;
 
 use Cicada\Core\Framework\App\AppEntity;
-use Cicada\Core\Framework\App\Event\AppDeactivatedEvent;
+use Cicada\Core\Framework\App\Event\AppActivatedEvent;
 use Cicada\Core\Framework\Context;
 use Cicada\Core\Framework\Uuid\Uuid;
 use Cicada\Core\Framework\Webhook\AclPrivilegeCollection;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
  */
-class AppDeactivatedEventTest extends TestCase
+#[CoversClass(AppActivatedEvent::class)]
+class AppActivatedEventTest extends TestCase
 {
     public function testGetter(): void
     {
         $app = new AppEntity();
         $context = Context::createDefaultContext();
-        $event = new AppDeactivatedEvent(
+        $event = new AppActivatedEvent(
             $app,
             $context
         );
 
         static::assertEquals($app, $event->getApp());
         static::assertEquals($context, $event->getContext());
-        static::assertSame(AppDeactivatedEvent::NAME, $event->getName());
+        static::assertSame(AppActivatedEvent::NAME, $event->getName());
         static::assertSame([], $event->getWebhookPayload());
     }
 
@@ -35,7 +37,7 @@ class AppDeactivatedEventTest extends TestCase
         $app = (new AppEntity())
             ->assign(['id' => $appId]);
         $context = Context::createDefaultContext();
-        $event = new AppDeactivatedEvent(
+        $event = new AppActivatedEvent(
             $app,
             $context
         );
