+++ tests/unit/Core/Framework/App/Event/AppDeletedEventTest.php
@@ -1,16 +1,18 @@
 <?php declare(strict_types=1);
 
-namespace Cicada\Tests\Integration\Core\Framework\App\Event;
+namespace Cicada\Tests\Unit\Core\Framework\App\Event;
 
 use Cicada\Core\Framework\App\Event\AppDeletedEvent;
 use Cicada\Core\Framework\Context;
 use Cicada\Core\Framework\Uuid\Uuid;
 use Cicada\Core\Framework\Webhook\AclPrivilegeCollection;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
  */
+#[CoversClass(AppDeletedEvent::class)]
 class AppDeletedEventTest extends TestCase
 {
     public function testGetter(): void
