+++ tests/unit/Core/Framework/App/Event/AppFlowActionEventTest.php
@@ -1,14 +1,16 @@
 <?php declare(strict_types=1);
 
-namespace Cicada\Tests\Integration\Core\Framework\App\Event;
+namespace Cicada\Tests\Unit\Core\Framework\App\Event;
 
 use Cicada\Core\Framework\App\Event\AppFlowActionEvent;
 use Cicada\Core\Framework\Webhook\AclPrivilegeCollection;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
  */
+#[CoversClass(AppFlowActionEvent::class)]
 class AppFlowActionEventTest extends TestCase
 {
     public function testGetter(): void
