+++ tests/unit/Core/Framework/FrameworkTest.php
@@ -38,6 +38,7 @@ class FrameworkTest extends TestCase
         $container->setParameter('cicada.cache.cache_compression', true);
         $container->setParameter('cicada.cache.cache_compression_method', 'gzip');
         $container->setParameter('kernel.debug', true);
+        $container->setParameter('kernel.environment', 'test');
         $framework = new Framework();
         $framework->setContainer($container);
